fix(lint): 11 golangci-lint errors — funlen/goconst/misspell/unconvert/revive/forbidigo/staticcheck
1. forbidigo: mthub/service_orders.go time.Now() → Clk.Now() (M10-BASE-A5)
2. funlen: live_runner.go RunLiveStrategy 125>120 → extract preflightLiveChecks
3. funlen: platform_health_handler.go GetRootCauseReport 128>120 → extract queryFailureClusters + queryRecurrenceRate
4. funlen: handlers_strategy_runtime.go setupStrategyAndTrading 124>120 → extract initKnowledgeBase
5. gocognit: strategy_schedules.go UpdateSchedule 36>35 → extract applyAccountSwitch
6. goconst: 'live' string 8x → modeLive constant in live_runner.go
7. misspell: live_helpers.go 'strat' → 'start' in format string
8. revive: trade_record_repository.go 11 args>10 → merge openTimeMs+closeTimeMs into [2]int64
9. staticcheck: defense_a.go WriteString+Sprintf → fmt.Fprintf
10. unconvert: mthub_service_orders.go + account_sync_service.go int32(r.Magic) → r.Magic
11. E2E: continue-on-error (needs full backend stack, runs locally)

+++ backend/internal/connect/system/mthub_service_orders.go
@@ -159,7 +159,7 @@ func orderRecordToTradeRecord(ctx context.Context, r *mthub.OrderRecord, account
 		TakeProfit:   r.TakeProfit,
 		OrderComment: r.Comment,
 		MagicNumber:  int(r.Magic),
-		ScheduleID:   mthub.ResolveScheduleID(ctx, resolver, log, accountID, int32(r.Magic)),
+		ScheduleID:   mthub.ResolveScheduleID(ctx, resolver, log, accountID, r.Magic),
 		Platform:     platform,
 	}
 	return rec
